+++ cdcv2/metadata/sql/client_orm.go
@@ -1,0 +1,456 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sql
+
+import (
+	"context"
+	"time"
+
+	"github.com/pingcap/log"
+	"github.com/pingcap/tiflow/cdc/model"
+	"github.com/pingcap/tiflow/pkg/errors"
+	"go.uber.org/zap"
+	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
+)
+
+var (
+	_ Checker[*gorm.DB] = &ormClient{}
+
+	_ upstreamClient[*gorm.DB]        = &ormClient{}
+	_ changefeedInfoClient[*gorm.DB]  = &ormClient{}
+	_ changefeedStateClient[*gorm.DB] = &ormClient{}
+	_ scheduleClient[*gorm.DB]        = &ormClient{}
+	_ progressClient[*gorm.DB]        = &ormClient{}
+)
+
+type ormClient struct {
+	selfID model.CaptureID
+	db     *gorm.DB
+}
+
+func NewORMClient(selfID model.CaptureID, db *gorm.DB) *ormClient {
+	return &ormClient{
+		selfID: selfID,
+		db:     db,
+	}
+}
+
+// Txn executes the given transaction action in a transaction.
+func (c *ormClient) Txn(ctx context.Context, fn ormTxnAction) error {
+	return c.db.WithContext(ctx).Transaction(fn)
+}
+
+// TxnWithOwnerLock executes the given transaction action in a transaction with owner lock.
+func (c *ormClient) TxnWithOwnerLock(ctx context.Context, uuid uint64, fn ormTxnAction) error {
+	return c.db.WithContext(ctx).Transaction(func(tx *gorm.DB) error {
+		var pr ProgressDO
+		ret := tx.Select("owner").
+			// TODO(charledCheung): use a variable to replace the hard-coded owner state.
+			Where("changefeed_uuid = ? and owner = ? and owner_state != removed", uuid, c.selfID).
+			Clauses(clause.Locking{
+				Strength: "SHARE",
+				Table:    clause.Table{Name: clause.CurrentTable},
+			}).Limit(1).Find(&pr)
+		if err := handleSingleOpErr(ret, 1, "TxnWithOwnerLock"); err != nil {
+			return errors.Trace(err)
+		}
+		return fn(tx)
+	})
+}
+
+// ================================ Upstream Client =================================
+
+// createUpstream implements the upstreamClient interface.
+func (s *ormClient) createUpstream(tx *gorm.DB, up *UpstreamDO) error {
+	ret := tx.Create(up)
+	if err := handleSingleOpErr(ret, 1, "CreateUpstream"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// deleteUpstream implements the upstreamClient interface.
+func (s *ormClient) deleteUpstream(tx *gorm.DB, up *UpstreamDO) error {
+	ret := tx.Delete(up)
+	if err := handleSingleOpErr(ret, 1, "DeleteUpstream"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// updateUpstream implements the upstreamClient interface.
+func (s *ormClient) updateUpstream(tx *gorm.DB, up *UpstreamDO) error {
+	ret := tx.Where("id = ? and version = ?", up.ID, up.Version).
+		Updates(UpstreamDO{
+			Endpoints: up.Endpoints,
+			Config:    up.Config,
+			Version:   up.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "UpdateUpstream"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// queryUpstreams implements the upstreamClient interface.
+func (s *ormClient) queryUpstreams(tx *gorm.DB) ([]*UpstreamDO, error) {
+	var ups []*UpstreamDO
+	ret := tx.Find(&ups)
+	if err := handleSingleOpErr(ret, -1, "QueryUpstreams"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return ups, nil
+}
+
+// queryUpstreamsByUpdateAt implements the upstreamClient interface.
+func (s *ormClient) queryUpstreamsByUpdateAt(tx *gorm.DB, lastUpdateAt time.Time) ([]*UpstreamDO, error) {
+	var ups []*UpstreamDO
+	ret := tx.Where("updated_at > ?", lastUpdateAt).Find(&ups)
+	if err := handleSingleOpErr(ret, -1, "QueryUpstreamsByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return ups, nil
+}
+
+// queryUpstreamByID implements the upstreamClient interface.
+func (s *ormClient) queryUpstreamByID(tx *gorm.DB, id uint64) (*UpstreamDO, error) {
+	var up *UpstreamDO
+	ret := tx.Where("id = ?", id).First(up)
+	if err := handleSingleOpErr(ret, 1, "QueryUpstreamsByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return up, nil
+}
+
+// ================================ ChangefeedInfo Client =================================
+
+// createChangefeedInfo implements the changefeedInfoClient interface.
+func (s *ormClient) createChangefeedInfo(tx *gorm.DB, info *ChangefeedInfoDO) error {
+	ret := tx.Create(info)
+	if err := handleSingleOpErr(ret, 1, "CreateChangefeedInfo"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// deleteChangefeedInfo implements the changefeedInfoClient interface.
+func (s *ormClient) deleteChangefeedInfo(tx *gorm.DB, info *ChangefeedInfoDO) error {
+	ret := tx.Delete(info)
+	if err := handleSingleOpErr(ret, 1, "DeleteChangefeedInfo"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// softDeleteChangefeedInfo implements the changefeedInfoClient interface.
+func (s *ormClient) softDeleteChangefeedInfo(tx *gorm.DB, info *ChangefeedInfoDO) error {
+	removeTime := time.Now()
+	ret := tx.Where("uuid = ? and version = ?", info.UUID, info.Version).
+		Updates(ChangefeedInfoDO{
+			RemovedAt: &removeTime,
+			Version:   info.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "SoftDeleteChangefeedInfo"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// updateChangefeedInfo implements the changefeedInfoClient interface.
+func (s *ormClient) updateChangefeedInfo(tx *gorm.DB, info *ChangefeedInfoDO) error {
+	ret := tx.Where("uuid = ? and version = ?", info.UUID, info.Version).
+		Updates(ChangefeedInfoDO{
+			SinkURI:  info.SinkURI,
+			StartTs:  info.StartTs,
+			TargetTs: info.TargetTs,
+			Config:   info.Config,
+			Version:  info.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "UpdateChangefeedInfo"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// queryChangefeedInfos implements the changefeedInfoClient interface.
+func (s *ormClient) queryChangefeedInfos(tx *gorm.DB) ([]*ChangefeedInfoDO, error) {
+	var infos []*ChangefeedInfoDO
+	ret := tx.Find(&infos)
+	if err := handleSingleOpErr(ret, -1, "QueryChangefeedInfos"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return infos, nil
+}
+
+// queryChangefeedInfosByUpdateAt implements the changefeedInfoClient interface.
+// TODO(CharlesCheung): query data before lastUpdateAt to avoid data loss.
+func (s *ormClient) queryChangefeedInfosByUpdateAt(tx *gorm.DB, lastUpdateAt time.Time) ([]*ChangefeedInfoDO, error) {
+	var infos []*ChangefeedInfoDO
+	ret := tx.Where("updated_at > ?", lastUpdateAt).Find(&infos)
+	if err := handleSingleOpErr(ret, -1, "QueryChangefeedInfosByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return infos, nil
+}
+
+// queryChangefeedInfoByUUID implements the changefeedInfoClient interface.
+func (s *ormClient) queryChangefeedInfoByUUID(tx *gorm.DB, uuid uint64) (*ChangefeedInfoDO, error) {
+	var info *ChangefeedInfoDO
+	ret := tx.Where("uuid = ?", uuid).First(info)
+
+	// TODO(CharlesCheung): handle record not found error.
+	if err := handleSingleOpErr(ret, 1, "QueryChangefeedInfoByUUID"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return info, nil
+}
+
+// ================================ ChangefeedState Client =================================
+
+// createChangefeedState implements the changefeedStateClient interface.
+func (s *ormClient) createChangefeedState(tx *gorm.DB, state *ChangefeedStateDO) error {
+	ret := tx.Create(state)
+	if err := handleSingleOpErr(ret, 1, "CreateChangefeedState"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// deleteChangefeedState implements the changefeedStateClient interface.
+func (s *ormClient) deleteChangefeedState(tx *gorm.DB, state *ChangefeedStateDO) error {
+	ret := tx.Delete(state)
+	if err := handleSingleOpErr(ret, 1, "DeleteChangefeedState"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// changefeedStateModel is used to update changefeed state, which prevents
+// nil values (error and warning) from being ignored.
+var changefeedStateModel = &ChangefeedStateDO{}
+
+// updateChangefeedState implements the changefeedStateClient interface.
+func (s *ormClient) updateChangefeedState(tx *gorm.DB, state *ChangefeedStateDO) error {
+	ret := tx.Model(changefeedStateModel).Select("state", "warning", "error", "version").
+		Where("changefeed_uuid = ? and version = ?", state.ChangefeedUUID, state.Version).
+		Updates(ChangefeedStateDO{
+			State:   state.State,
+			Warning: state.Warning,
+			Error:   state.Error,
+			Version: state.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "updateChangefeedState"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// queryChangefeedStates implements the changefeedStateClient interface.
+func (s *ormClient) queryChangefeedStates(tx *gorm.DB) ([]*ChangefeedStateDO, error) {
+	var states []*ChangefeedStateDO
+	ret := tx.Find(&states)
+	if err := handleSingleOpErr(ret, -1, "QueryChangefeedStates"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return states, nil
+}
+
+// queryChangefeedStatesByUpdateAt implements the changefeedStateClient interface.
+func (s *ormClient) queryChangefeedStatesByUpdateAt(tx *gorm.DB, lastUpdateAt time.Time) ([]*ChangefeedStateDO, error) {
+	var states []*ChangefeedStateDO
+	ret := tx.Where("updated_at > ?", lastUpdateAt).Find(&states)
+	if err := handleSingleOpErr(ret, -1, "QueryChangefeedStatesByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return states, nil
+}
+
+// queryChangefeedStateByUUID implements the changefeedStateClient interface.
+func (s *ormClient) queryChangefeedStateByUUID(tx *gorm.DB, uuid uint64) (*ChangefeedStateDO, error) {
+	var state *ChangefeedStateDO
+	ret := tx.Where("changefeed_uuid = ?", uuid).First(state)
+	if err := handleSingleOpErr(ret, 1, "QueryChangefeedStateByUUID"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return state, nil
+}
+
+// ================================ Schedule Client =================================
+
+// createSchedule implements the scheduleClient interface.
+func (s *ormClient) createSchedule(tx *gorm.DB, sc *ScheduleDO) error {
+	ret := tx.Create(sc)
+	if err := handleSingleOpErr(ret, 1, "CreateSchedule"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// deleteSchedule implements the scheduleClient interface.
+func (s *ormClient) deleteSchedule(tx *gorm.DB, sc *ScheduleDO) error {
+	ret := tx.Delete(sc)
+	if err := handleSingleOpErr(ret, 1, "DeleteSchedule"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// updateSchedule implements the scheduleClient interface.
+func (s *ormClient) updateSchedule(tx *gorm.DB, sc *ScheduleDO) error {
+	ret := tx.Where("changefeed_uuid = ? and version = ?", sc.ChangefeedUUID, sc.Version).
+		Updates(ScheduleDO{
+			Owner:        sc.Owner,
+			OwnerState:   sc.OwnerState,
+			Processors:   sc.Processors,
+			TaskPosition: sc.TaskPosition,
+			Version:      sc.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "UpdateSchedule"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// querySchedules implements the scheduleClient interface.
+func (s *ormClient) querySchedules(tx *gorm.DB) ([]*ScheduleDO, error) {
+	var schedules []*ScheduleDO
+	ret := tx.Find(&schedules)
+	if err := handleSingleOpErr(ret, -1, "QuerySchedules"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return schedules, nil
+}
+
+// querySchedulesByUpdateAt implements the scheduleClient interface.
+func (s *ormClient) querySchedulesByUpdateAt(tx *gorm.DB, lastUpdateAt time.Time) ([]*ScheduleDO, error) {
+	var schedules []*ScheduleDO
+	ret := tx.Where("updated_at > ?", lastUpdateAt).Find(&schedules)
+	if err := handleSingleOpErr(ret, -1, "QuerySchedulesByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return schedules, nil
+}
+
+// querySchedulesByOwnerIDAndUpdateAt implements the scheduleClient interface.
+func (s *ormClient) querySchedulesByOwnerIDAndUpdateAt(tx *gorm.DB, captureID string, lastUpdateAt time.Time) ([]*ScheduleDO, error) {
+	var schedules []*ScheduleDO
+	ret := tx.Where("owner = ? and updated_at > ?", captureID, lastUpdateAt).Find(&schedules)
+	if err := handleSingleOpErr(ret, -1, "QuerySchedulesByOwnerIDAndUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return schedules, nil
+}
+
+// queryScheduleByUUID implements the scheduleClient interface.
+func (s *ormClient) queryScheduleByUUID(tx *gorm.DB, uuid uint64) (*ScheduleDO, error) {
+	var schedule *ScheduleDO
+	ret := tx.Where("changefeed_uuid = ?", uuid).First(schedule)
+	if err := handleSingleOpErr(ret, 1, "QueryScheduleByUUID"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return schedule, nil
+}
+
+// ================================ Progress Client =================================
+
+// createProgress implements the progressClient interface.
+func (s *ormClient) createProgress(tx *gorm.DB, pr *ProgressDO) error {
+	ret := tx.Create(pr)
+	if err := handleSingleOpErr(ret, 1, "CreateProgress"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// deleteProgress implements the progressClient interface.
+func (s *ormClient) deleteProgress(tx *gorm.DB, pr *ProgressDO) error {
+	ret := tx.Delete(pr)
+	if err := handleSingleOpErr(ret, 1, "DeleteProgress"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// updateProgress implements the progressClient interface.
+func (s *ormClient) updateProgress(tx *gorm.DB, pr *ProgressDO) error {
+	ret := tx.Where("capture_id = ? and version = ?", pr.CaptureID, pr.Version).
+		Updates(ProgressDO{
+			Progress: pr.Progress,
+			Version:  pr.Version + 1,
+		})
+	if err := handleSingleOpErr(ret, 1, "UpdateProgress"); err != nil {
+		return errors.Trace(err)
+	}
+	return nil
+}
+
+// queryProgresss implements the progressClient interface.
+func (s *ormClient) queryProgresss(tx *gorm.DB) ([]*ProgressDO, error) {
+	var progresses []*ProgressDO
+	ret := tx.Find(&progresses)
+	if err := handleSingleOpErr(ret, -1, "QueryProgresss"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return progresses, nil
+}
+
+// queryProgresssByUpdateAt implements the progressClient interface.
+func (s *ormClient) queryProgresssByUpdateAt(tx *gorm.DB, lastUpdateAt time.Time) ([]*ProgressDO, error) {
+	var progresses []*ProgressDO
+	ret := tx.Where("updated_at > ?", lastUpdateAt).Find(&progresses)
+	if err := handleSingleOpErr(ret, -1, "QueryProgresssByUpdateAt"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return progresses, nil
+}
+
+// queryProgressByCaptureID implements the progressClient interface.
+func (s *ormClient) queryProgressByCaptureID(tx *gorm.DB, id string) (*ProgressDO, error) {
+	var progress *ProgressDO
+	ret := tx.Where("capture_id = ?", id).First(progress)
+	if err := handleSingleOpErr(ret, 1, "QueryProgressByCaptureID"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return progress, nil
+}
+
+// queryProgressByCaptureIDsWithLock implements the progressClient interface.
+func (s *ormClient) queryProgressByCaptureIDsWithLock(tx *gorm.DB, ids []string) ([]*ProgressDO, error) {
+	var progresses []*ProgressDO
+	ret := tx.Where("capture_id in (?)", ids).Clauses(clause.Locking{
+		Strength: "SHARE",
+		Table:    clause.Table{Name: clause.CurrentTable},
+	}).Find(&progresses)
+	if err := handleSingleOpErr(ret, -1, "QueryProgressByCaptureIDsWithLock"); err != nil {
+		return nil, errors.Trace(err)
+	}
+	return progresses, nil
+}
+
+func handleSingleOpErr(ret *gorm.DB, targetRows int64, opName string) (err error) {
+	defer func() {
+		if err != nil {
+			log.Error("run unary meta operation failed", zap.Error(err))
+		}
+	}()
+
+	if ret.Error != nil {
+		return errors.WrapError(errors.ErrMetaOpFailed, ret.Error, opName)
+	}
+	if targetRows >= 0 && ret.RowsAffected != targetRows {
+		return errors.ErrMetaRowsAffectedNotMatch.GenWithStackByArgs(opName, targetRows, ret.RowsAffected)
+	}
+	return nil
+}
